Renamed old test folder and added a few tests 🙆
