+++ package.json
@@ -8,7 +8,7 @@
     "lannister": "./bin/lannister.js"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha"
   },
   "keywords": [
     "technical",
@@ -17,9 +17,12 @@
   "author": "Kenneth Larsen",
   "license": "ISC",
   "devDependencies": {
+    "chai": "^4.1.2",
     "eslint": "^4.9.0",
     "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-plugin-import": "^2.7.0"
+    "eslint-plugin-import": "^2.7.0",
+    "mocha": "^5.0.5",
+    "mock-fs": "^4.4.2"
   },
   "dependencies": {
     "file-extension": "^4.0.2",
